add support for used memory
